refactor (rename + move) JavaFXThreadingRule
- Move to Gui package from test package
- Rename from JavaFXThreadingRule to JavaFXInitTestRule
